Stats system
RequestService has a general method to post to endpoint with body
QuizView saves the currentQuestion for QuizController.next()
Stats model attribute "order" is changed to "index"
Quiz Model has function addStats, which also auto generates the current UNIX Epoch in int.
StatsDAO has the function to post Stats object to the API by converting the model into a JsonObject.
QuizController can decides when a stat gets added and when everything is posted to the API.